server/requirements.txt
# Core FastAPI and server
fastapi[all]>=0.104.0
uvicorn[standard]>=0.24.0

# Database and Supabase
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP client with socks support (fixes socksio error)
httpx[socks]>=0.25.0
socksio>=1.0.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Email sending
aiosmtplib>=2.0.0
email-validator>=2.1.0

# Scheduling and background tasks
apscheduler>=3.10.0

# Push notifications (Expo)
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
pytz>=2023.3
dateutils>=0.6.12
validators>=0.22.0